requirements.txt
py-cord>=2.4.0
python-dotenv>=1.0.0
trueskill>=0.4.5